refactor(desktop): 提取扩展与市场协调逻辑
